m55736 LoD support (2)

+++ source/Renderer/src/ViewWeightingSynthesizer.cpp
@@ -508,9 +508,9 @@ private:
 
           for (auto y = y0; y < y1; y++) {
             for (auto x = x0; x < x1; x++) {
-              // temporary use only view dimensions
-              if (y >= m_sourceReprojection[viewId].height() ||
-                  x >= m_sourceReprojection[viewId].width()) {
+              // TODO(BK): Can this be removed?
+              if (static_cast<uint32_t>(y) >= m_sourceReprojection[viewId].height() ||
+                  static_cast<uint32_t>(x) >= m_sourceReprojection[viewId].width()) {
                 continue;
               }
